refactor: component 에서 css style component 분리
