+++ src/components/@common/bottomModal/BottomModal.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import styled from 'styled-components';
+import { StyleBackDrop, StyleContainer } from './BottomModal.style';
 
 interface Props extends React.PropsWithChildren {
   isOpen: boolean;
@@ -22,41 +22,15 @@ const BottomModal = ({ isOpen, onClose, children }: Props) => {
 
   return isOpen ? (
     <>
-      <Backdrop
+      <StyleBackDrop
         className="modal-backdrop"
         onClick={() => {
           onClose();
         }}
-      ></Backdrop>
-      <Container>{children}</Container>
+      ></StyleBackDrop>
+      <StyleContainer>{children}</StyleContainer>
     </>
   ) : null;
 };
-
-const Backdrop = styled.div`
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-
-  background: rgba(0, 0, 0, 0.35);
-`;
-
-const Container = styled.div`
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-
-  box-sizing: border-box;
-
-  padding: 32px 16px;
-
-  border: 0;
-
-  border-radius: 8px 8px 0px 0px;
-  background: white;
-`;
 
 export default BottomModal;
