+++ src/components/cardCompanyIconBox/CardCompanyIconBox.style.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const ImageContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const StyledBankImage = styled.img`
+  width: 37px;
+  height: 37px;
+  margin: auto;
+`;
