+++ src/components/registerForm/cardNumber/CardNumber.tsx
@@ -2,7 +2,7 @@ import React, { useState, useContext } from 'react';
 
 import useInputListRef from '../../../hooks/useInputListRef';
 import FormLabel from '../../@common/FormLabel';
-import Input from '../../@common/Input';
+import Input from '../../@common/input/Input';
 import ErrorSpan from '../../@common/ErrorSpan';
 import InputWrapper from '../../@common/InputWrapper';
 import CardNumberInputContainer from './CardNumber.style';
